+++ .gitignore
@@ -1,4 +1,5 @@
 venv
 *.pyc
 .idea/
-
+*.html
+assets
