+++ Pages/Page_Login.py
@@ -6,7 +6,7 @@ class LoginPage:
     password = (By.ID, "password")
     login_button = (By.ID, "login-button")
     product = (By.CLASS_NAME, "title")
-    error_msg = (By.CLASS_NAME, "error-message-container.error")
+    error_msg = (By.CSS_SELECTOR, ".error-message-container.error")
     error_close_btn = (By.CLASS_NAME, "error-button")
