+++ tests/test_products.py
@@ -1,9 +1,20 @@
+import pytest
 from selenium.webdriver.support.wait import WebDriverWait
 from selenium.webdriver.support import expected_conditions as EC
+
+from Pages.Page_Login import LoginPage
 from Pages.Page_MenuBar import MenuBarPage
 from Pages.Page_Inventory import InventoryPage
 from utility.utilities import getdata
 
+
+@pytest.fixture(scope="function", autouse=True)
+def login(driver):
+    driver.find_element(*LoginPage.user_name).click()
+    driver.find_element(*LoginPage.user_name).send_keys(getdata("user1"))
+    driver.find_element(*LoginPage.password).click()
+    driver.find_element(*LoginPage.password).send_keys(getdata("password"))
+    driver.find_element(*LoginPage.login_button).click()
 
 def test_menu(driver, login):
     driver.find_element(*MenuBarPage.menu_bar).click()
@@ -16,17 +27,17 @@ def test_menu(driver, login):
     assert menu_options == getdata("menu-items"), "Expected menu options are not available"
 
 
-def test_add_to_cart(driver, login):
-    driver.find_element(*InventoryPage.bag_pack_add).click()
-    count = driver.find_element(*InventoryPage.shoppingcart).text
-    assert int(count) == 1, "After adding item, cart doesn't show correct number"
+# def test_add_to_cart(driver, login):
+#     driver.find_element(*InventoryPage.bag_pack_add).click()
+#     count = driver.find_element(*InventoryPage.shoppingcart).text
+#     assert int(count) == 1, "After adding item, cart doesn't show correct number"
 
 
-def test_remove_from_cart(driver, login):
-    driver.find_element(*InventoryPage.bag_pack_add).click()
+def test_add_and_remove_from_cart(driver, login):
+    driver.find_element(*InventoryPage.add_to_cart_button).click()
     count = driver.find_element(*InventoryPage.shoppingcart).text
     assert count == '1', "After adding item, cart doesn't show correct number"
-    driver.find_element(*InventoryPage.bag_pack_remove).click()
+    driver.find_element(*InventoryPage.remove_button).click()
     new_count = driver.find_element(*InventoryPage.shoppingcart).text
     assert new_count == '', "After removing item, cart doesn't show correct number"
 
@@ -47,10 +58,10 @@ def test_product_details(driver, login):
 
 def test_product_detail_page(driver, login):
     driver.find_element(*InventoryPage.bag_pack_detail).click()
+    WebDriverWait(driver, 10).until(EC.visibility_of_all_elements_located(InventoryPage.back_to_products_btn))
     assert driver.find_element(*InventoryPage.back_to_products_btn).is_displayed(), "Backtoproduct btn is not available"
-    assert driver.find_element(*InventoryPage.bag_pack_add).is_displayed(), "Add to cart button is not available"
-    driver.find_element(*InventoryPage.bag_pack_add).click()
-    assert driver.find_element(*InventoryPage.bag_pack_remove).is_displayed(), "Remove button is not available"
-
-def test_1():
-    pass
+    assert driver.find_element(*InventoryPage.add_to_cart_button).is_displayed(), "Add to cart button is not available"
+    WebDriverWait(driver,10).until(EC.element_to_be_clickable(InventoryPage.add_to_cart_button))
+    driver.find_element(*InventoryPage.add_to_cart_button).click()
+    # WebDriverWait(driver, 10).until(EC.visibility_of_all_elements_located(InventoryPage.bag_pack_remove))
+    assert driver.find_element(*InventoryPage.remove_button).is_displayed(), "Remove button is not available"
